requirements.txt
# process images into matrrices for Tensorflow
numpy==1.11.0
Pillow==3.1.1
jupyter==1.0.0

# download images from S3
boto3==1.3.0

GDAL==2.0.1
pyproj==1.9.5.1

# the recommended Tensorflow binary on linux, CPU only
https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow-0.8.0-cp27-none-linux_x86_64.whl

# the recommended Tensorflow binary for python 2 on linux, GPU
# http://storage.googleapis.com/tensorflow/linux/gpu/tensorflow-0.8.0-cp27-none-linux_x86_64.whl

shapely

# trying this out for the cifar10 network
git+https://github.com/tflearn/tflearn.git

h5py